requirements.txt
streamlit>=1.33.0
folium>=0.16.0
streamlit-folium>=0.15.0
python-gedcom>=0.1.0
pandas>=2.2.2
protobuf>=6.30.0,<7.0dev
